Results and source from version 2 of testing.
A lot of the intent has changed from this little experiment. I changed
from trying to figure out threading in java and python to figuring out
the differences in speed between already build and actively maintained
libraries and my attempt at a speedy algorithm.

This commit includes these files:
 * opencv_avg.py - OpenCV implementation of mean() in python.
 * AverageColor.cpp - OpenCV implementation of mean() in C++.
 * QuickForLoopAvg.java - A bit/bytewise access implementation in Java.
 * ThreadedAvg.java - An inefficient threaded implementation in Java
(included here to maybe come back to it later and for safe keeping).
 * avgtime.sh - The timing script that will have its own repo soon.